Create sorthem.asm
This is an assembly application I built in NASM for a course at McMaster University.  The program requires one integer argument between 1 and 9.  Then the program displays a stack of discs in a random order corresponding to the number entered by the user. The program then displays the sorting of the stack.